requirements.txt
requests>=2.28.2
pandas>=1.5.3
python-dotenv>=1.0.0
ratelimit>=2.2.1
python-dateutil>=2.8.2
pydantic>=1.10.7
sqlalchemy>=1.4.46
alembic>=1.10.2
pytest>=7.4.2
